Installed BackUpWordPress
BackUpWordPress appears third in the plugintable.com listing for backup
plugins at almost 200,000 downloads as of end August 2015. The goal to
including plugins is to have a well known free version with a paid
upgrade that is easy to configure and comes with default settings in
place on activation. This one does that. This means that, along with the
Sucuri Security plugin, WP Super Cache, Autoptimize, my own flavour of
Optimization settings, as well as an Installer plugin, you should have a
decent bundled set to use as a starting point for further development.
Backed up, secured, cached and optimized. A decent starting point for a
WordPress site.