update
(re-installed backupwordpress by Human Made Ltd.). It is a great plugin
that automatically configures itself on activation, but it would be nice
if it loaded files only when needed, as a performance check showed it
created more drag than it should need to. It shouldn't need to load much
of anything on each page load, except when displaying it's page in the
admin section.